backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-core
langchain-experimental
langchain-neo4j
langchain-google-genai
neo4j
tqdm-loggable
python-dotenv
tiktoken
wikipedia
unstructured
markdown
yfiles-jupyter-graphs
asyncio